defined cleaner functions and created app menu

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import create_engine, Column, Date, Integer, String, ForeignKey
+from sqlalchemy import create_engine, Column, Date, Integer, String, Float, ForeignKey
 from sqlalchemy.orm import sessionmaker, declarative_base
 
 engine = create_engine('sqlite:///inventory.db', echo=False)
@@ -16,6 +16,6 @@ class Product(Base):
     product_id = Column(Integer, primary_key=True)
     product_name = Column(String)
     product_quantity = Column(Integer)
-    product_price = Column(Integer)
+    product_price = Column(Float)
     date_updated = Column(Date)
     brand_id = Column(Integer, ForeignKey('brands.brand_id'))
